Changed: Split reusable components from calc_compression_stats into separate files
With some extra cleanup
